+++ cjs/index.d.ts
@@ -1,8 +1,12 @@
 /**
  * Computes and returns the factorial of `n`.
- * Returns -1 if `n` is negative, or `NaN` if it's infinite, itself `NaN` or a float.
+ * Returns `NaN` if `n` is negative, infinite, itself `NaN` or a float.
  *
  * @param n The number for which the factorial is to be calculated.
  * @returns `n!`
  */
-export declare function factorial(n: number): number;
+declare function factorial(n: number): number;
+declare namespace factorial {
+    var factorial: typeof import(".");
+}
+export = factorial;