+++ index.js
@@ -1,3 +1,0 @@
-import def from './cjs/index.js';
-
-export const factorial = def.factorial;