+++ src/index.ts
@@ -1,28 +1,18 @@
-function __checkErrors(n: number) {
-	if (typeof n !== 'number')
-		throw new TypeError(
-			`"n" must be a number. Got "${n}" of type "${typeof n}".`
-		);
-}
-function __isInteger(num: number) {
-	return (
-		Number?.isInteger?.(num) ||
-		(num < 0 ? Math.ceil(num) : Math.floor(num)) === num
-	);
-}
+import assertTypeOf = require('@santi100/assertion-lib/cjs/type-of');
+
 /**
  * Computes and returns the factorial of `n`.
- * Returns -1 if `n` is negative, or `NaN` if it's infinite, itself `NaN` or a float.
+ * Returns `NaN` if `n` is negative, infinite, itself `NaN` or a float.
  *
  * @param n The number for which the factorial is to be calculated.
  * @returns `n!`
  */
-export function factorial(n: number) {
-	__checkErrors(n);
-	if (isNaN(n) || !__isInteger(n) || n >= Infinity || n <= -Infinity)
-		return NaN;
+function factorial(n: number) {
+	assertTypeOf(n, 'number', 'n');
 	let fact = 1;
-	if (n < 0) return -1;
+	if (isNaN(n) || (n - Math.floor(n) !== 0) || n >= Infinity || n <= -Infinity)
+		return NaN;
+	if (n < 0) return NaN;
 	else {
 		for (let i = 1; i <= n; i++) {
 			fact *= i;
@@ -30,3 +20,9 @@ export function factorial(n: number) {
 	}
 	return fact;
 }
+
+factorial.factorial = factorial;
+
+Object?.defineProperty?.(factorial, 'factorial', { enumerable: false });
+
+export = factorial;