+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 
 ## Version 0.0.1
+
 - First version!
 
 ## Version 0.0.2
-- Added heap sort.
+
+- Modified dependencies.
+- Modified `factorial` to return `NaN` instead of -1 when the input is smaller than zero.